requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-extensions==3.2.3
django-filter==24.2
django-sslserver==0.22
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
idna==3.7
libretranslatepy==2.1.1
lxml==5.2.2
Markdown==3.6
MarkupSafe==2.1.5
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.82
packaging==24.1
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
pyOpenSSL==24.1.0
pytesseract==0.3.10
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
translate==3.6.1
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
